refactor p2pk types from base
