+++ tests/test_wallet_p2pk.py
@@ -6,9 +6,10 @@ from typing import List
 import pytest
 import pytest_asyncio
 
-from cashu.core.base import Proof, SigFlags, Tags
+from cashu.core.base import Proof
 from cashu.core.crypto.secp import PrivateKey, PublicKey
 from cashu.core.migrations import migrate_databases
+from cashu.core.p2pk import SigFlags, Tags
 from cashu.wallet import migrations
 from cashu.wallet.wallet import Wallet
 from cashu.wallet.wallet import Wallet as Wallet1
